requirements.txt
streamlit==1.44.1
requests==2.32.3
pydantic==2.11.4
python-dotenv==1.0.1
python-docx==1.1.2
pypdf==4.3.1
